refactor(playground): split websocket connection hook
