+++ packages/agent-playground/src/hooks/use-websocket-connection/constants.ts
@@ -1,0 +1,8 @@
+export const DEFAULT_RECONNECT_INTERVAL_MS = 1000;
+export const DEFAULT_MAX_RECONNECT_ATTEMPTS = 5;
+export const UPTIME_UPDATE_INTERVAL_MS = 1000;
+export const MAX_RECONNECT_DELAY_MS = 30000;
+export const PING_TIMEOUT_MS = 5000;
+export const PING_MIN_LATENCY_MS = 50;
+export const PING_MAX_EXTRA_LATENCY_MS = 100;
+export const HIGH_CONNECTION_ATTEMPTS_THRESHOLD = 3;
