+++ packages/agent-playground/src/hooks/use-websocket-connection/index.ts
@@ -1,0 +1,9 @@
+export { useWebSocketConnection } from './use-websocket-connection';
+export type {
+  IConnectionHealth,
+  IConnectionInfo,
+  IConnectionStatistics,
+  IWebSocketAuth,
+  IWebSocketConnectionHookReturn,
+  TConnectionState,
+} from './types';
